refactor: Explicit node tracker import
